Refactor documentation structure and update project metadata
- Removed unnecessary subpackage documentation for CityOfBinds.utils and CityOfBinds.utils.types.
- Simplified the API reference sections in the HTML documentation.
- Updated the project metadata in conf.py to dynamically read author and version from the CityOfBinds package.
- Adjusted the index.rst to streamline the table of contents.
- Removed redundant sections in CityOfBinds.rst.
- Updated pyproject.toml to use dynamic versioning based on the package's __version__ attribute.
