+++ CityOfBinds/_version.py
@@ -1,0 +1,5 @@
+"""Version information for CityOfBinds."""
+
+__version__ = "0.0.1"
+__author__ = "brandon wade"
+__email__ = "bwade.dev@gmail.com"
