+++ CityOfBinds/src/game/rotating_binds/rotating_bind.py
@@ -1,4 +1,4 @@
-from ....utils.types import StrPath
+from ....utils.types.str_path import StrPath
 from .generic_rotating_binds import _GenericRotatingBind, _LoopTopology
