--- CityOfBinds/utils/types/__init__.py
@@ -1,3 +1,0 @@
-from .str_path import StrPath
-
-__all__ = ["StrPath"]
