--- docs/source/CityOfBinds.utils.rst
@@ -1,45 +1,0 @@
-CityOfBinds.utils package
-=========================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   CityOfBinds.utils.types
-
-Submodules
-----------
-
-CityOfBinds.utils.baseconverter module
---------------------------------------
-
-.. automodule:: CityOfBinds.utils.baseconverter
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-CityOfBinds.utils.file\_graph\_publisher module
------------------------------------------------
-
-.. automodule:: CityOfBinds.utils.file_graph_publisher
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-CityOfBinds.utils.pathgenerator module
---------------------------------------
-
-.. automodule:: CityOfBinds.utils.pathgenerator
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: CityOfBinds.utils
-   :members:
-   :show-inheritance:
-   :undoc-members:
