--- docs/source/CityOfBinds.utils.types.rst
@@ -1,21 +1,0 @@
-CityOfBinds.utils.types package
-===============================
-
-Submodules
-----------
-
-CityOfBinds.utils.types.str\_path module
-----------------------------------------
-
-.. automodule:: CityOfBinds.utils.types.str_path
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: CityOfBinds.utils.types
-   :members:
-   :show-inheritance:
-   :undoc-members:
